created different files for server and client side tests
